Reasses structure of training config and dataset creation logic
